change float to numeric

+++ ivy_tests/test_ivy/test_frontends/test_torch/test_miscellaneous_ops.py
@@ -141,7 +141,7 @@ def test_torch_roll(
 @handle_frontend_test(
     fn_tree="torch.meshgrid",
     dtypes_and_tensors=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("float"),
+        available_dtypes=helpers.get_dtypes("numeric"),
         num_arrays=st.integers(min_value=2, max_value=5),
         min_num_dims=1,
         max_num_dims=1,
